requirements.txt
# Core
pytz==2018.5
Django==2.1.1
django-configurations==2.1
gunicorn==19.9.0
newrelic==4.4.0.103
django-extensions

# For the persistence stores
dj-database-url==0.5.0
mysql

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1
django-modeltranslation==0.14.4

# Rest apis
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==2.0.0
drf_yasg==1.11.0
flex==6.13.2
swagger-spec-validator==2.4.1
django-cors-headers==2.4.0
xmltodict==0.12.0

# Celery
celery==3.1.18
redis==2.10.3
django-celery==3.2.2
tornado==4.5.3
flower==0.9.2

# Developer Tools
ipdb==0.11
ipython==6.5.0
mkdocs==0.17.5
flake8==3.5.0

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.5
nose-progressive==1.5.1
coverage==4.5.1

# Static and Media Storage
django-storages==1.7.1
boto3==1.9.4
whitenoise==4.1.2

# Internal library
-e git+https://erasmo_montes@bitbucket.org/grupopuntacana/partenon.git#egg=partenon
-e git+https://erasmo_montes@bitbucket.org/grupopuntacana/oraculo.git#egg=oraculo

